requirements.txt
pytest==8.3.2
requests==2.32.3
PyMySQL==1.1.1
elasticsearch==8.13.2
confluent-kafka==2.6.0
confluent-kafka[avro]==2.6.0
fastavro==1.9.7
PyYAML==6.0.1
mysql-connector-python==9.0.0

